add cp Question

+++ Competitive_programming/squere_problem.java
@@ -5,12 +5,12 @@ import java.util.Scanner;
 public class squere_problem {
     public static void main(String[] args) {
         Scanner sc=new Scanner(System.in);
-        int big_breath=sc.nextInt();
-        int big_height=sc.nextInt();
-        int small_breath=sc.nextInt();
-        int small_height=small_breath;
-        int b_num=0;
-        int h_num=0;
+        long big_breath=sc.nextInt();
+        long big_height=sc.nextInt();
+        long small_breath=sc.nextInt();
+        long small_height=small_breath;
+        long b_num=0;
+        long h_num=0;
 
         if(small_breath*small_height<=big_breath*big_height){
             if(big_breath%small_breath!=0)
